Move api client to api_client namespace
